Correctly implement regenerate halfMap when halfMap invalid
Removed HashMap<Position, EnumTerrain> nodes from the class fields because it was causing generateHalfMap() to stop on a second time as nodes was already populated with data (could have just set it with null in generateHalfMap()); instead, I pass it as parameter to all functions.

Added new unit tests, fixed small problems.